minor refactor of mix_fuel, so its more general and I can use it for WAR too
